# app/src/apiDef/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Online Shopping",
    "description": "This is a sample shopping server.\n",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "localhost:8080/moezguer/Shopping/v1",
      "description": "local virtual server"
    },
    {
      "url": "https://dev.swaggerhub.com/moezguer/Shopping/v1",
      "description": "Development server"
    },
    {
      "url": "https://stag.swaggerhub.com/moezguer/Shopping/v1",
      "description": "Staging server"
    },
    {
      "url": "https://prod.swaggerhub.com/moezguer/Shopping/v1",
      "description": "Production server"
    }
  ],
...
  "paths": {
    "/products": {
      "get": {
        "tags": [
          "products"
        ],
        "description": "Returns all the products",
        "operationId": "getAllProducts",
        "responses": {
          "200": {
            "description": "Successful product operations list response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Product"
                  }
                },
                "examples": {
                  "exProductResponseList": {
                    "value": [
                      {
                        "id": "883c19ea-443f-42b3-80a9-ae5b98968c15",
                        "title": "Hair Dryer",
                        "price": 11.44,
                        "currency": "TRY"
                      },
                      {
                        "id": "c2c97cdd-642c-4d42-87a3-5a31f6cee511",
                        "title": "Cleaner",
                        "price": 6.8,
                        "currency": "TRY"
                      }
                    ]
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "products"
        ],
        "description": "Adds product",
        "operationId": "addProduct",
        "requestBody": {
          "$ref": "#/components/requestBodies/productRequest"
        },
        "responses": {
          "201": {
            "description": "Successful product operations response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponse"
                },
                "examples": {
                  "exProductResponse": {
                    "value": {
                      "id": "883c19ea-443f-42b3-80a9-ae5b98968c15",
                      "title": "Hair Dryer",
                      "price": 11.44,
                      "currency": "TRY"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Request invalid"
          }
        }
      }
    },
    "/products/{productid}": {
      "get": {
        "tags": [
          "products"
        ],
        "description": "Return product with the ID",
        "operationId": "findProductById",
        "parameters": [
          {
            "name": "productid",
            "in": "path",
            "description": "Product identifier.",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "examples": {
              "exProductId": {
                "value": "883c19ea-443f-42b3-80a9-ae5b98968c15"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful product operations response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponse"
                },
                "examples": {
                  "exProductResponse": {
                    "value": {
                      "id": "883c19ea-443f-42b3-80a9-ae5b98968c15",
                      "title": "Hair Dryer",
                      "price": 11.44,
                      "currency": "TRY"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "ID not found"
          }
        }
      },
      "put": {
        "tags": [
          "products"
        ],
        "description": "Updates product",
        "operationId": "updateProducts",
        "parameters": [
          {
            "name": "productid",
            "in": "path",
            "description": "Product identifier.",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "examples": {
              "exProductId": {
                "value": "883c19ea-443f-42b3-80a9-ae5b98968c15"
              }
            }
          }
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/productRequest"
        },
        "responses": {
          "204": {
            "description": "Successful product operations response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponse"
                },
                "examples": {
                  "exProductResponse": {
                    "value": {
                      "id": "883c19ea-443f-42b3-80a9-ae5b98968c15",
                      "title": "Hair Dryer",
                      "price": 11.44,
                      "currency": "TRY"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Request invalid"
          },
          "404": {
            "description": "ID not found"
          }
        }
      },
      "delete": {
        "tags": [
          "products"
        ],
        "description": "Deletes product by ID",
        "operationId": "deleteProduct",
        "parameters": [
          {
            "name": "productid",
            "in": "path",
            "description": "Product identifier.",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "examples": {
              "exProductId": {
                "value": "883c19ea-443f-42b3-80a9-ae5b98968c15"
              }
            }
          }
        ],
        "responses": {
          "205": {
            "description": "OK"
          },
          "404": {
            "description": "ID not found"
          }
        }
      }
    },
    "/products/import/list": {
      "post": {
        "tags": [
          "products"
        ],
        "description": "imports products via list",
        "operationId": "importProductsViaList",
        "requestBody": {
          "$ref": "#/components/requestBodies/productRequestList"
        },
        "responses": {
          "201": {
            "description": "Successfully imported"
          },
          "400": {
            "description": "Request invalid"
          }
        }
      }
    },
    "/products/import/file/csv": {
      "post": {
        "tags": [
          "products"
        ],
        "description": "imports products via CSV file",
        "operationId": "importProductsViaFile",
        "requestBody": {
          "content": {
            "application/octet-stream": {
              "schema": {
                "type": "string",
                "format": "binary"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successfully imported"
          },
          "400": {
            "description": "Request invalid"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Currency": {
        "type": "string",
        "default": "TRY",
        "enum": [
          "TRY",
          "EUR",
          "USD"
        ]
      },
      "Product": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "maximum": 80,
            "minimum": 2,
            "type": "string"
          },
          "price": {
            "type": "number",
            "format": "double"
          },
          "currency": {
            "$ref": "#/components/schemas/Currency"
          }
        }
      },
      "ProductRequest": {
        "type": "object",
        "properties": {
          "title": {
            "maximum": 80,
            "minimum": 2,
            "type": "string"
          },
          "price": {
            "type": "number",
            "format": "double"
          },
          "currency": {
            "$ref": "#/components/schemas/Currency"
          }
        }
      },
      "ProductResponse": {
        "$ref": "#/components/schemas/Product"
      },
...
    },
...
    "requestBodies": {
      "productRequest": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ProductRequest"
            },
            "examples": {
              "exProductRequest": {
                "value": {
                  "title": "Hair Dryer",
                  "price": 11.44,
                  "currency": "TRY"
                }
              }
            }
          }
        },
        "required": true
      },
      "productRequestList": {
        "content": {
          "application/json": {
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/ProductRequest"
              }
            },
            "examples": {
              "exProductRequestList": {
                "value": [
                  {
                    "title": "Hair Dryer",
                    "price": 11.44,
                    "currency": "TRY"
                  },
                  {
                    "title": "Cleaner",
                    "price": 6.8,
                    "currency": "TRY"
                  }
                ]
              }
            }
          }
        },
        "required": true
      },
...
    }
  }
}