# app/src/main/java/io/swagger/dto/model/Product.java
package io.swagger.dto.model;

import io.swagger.enumeration.Currency;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.util.UUID;

@Data
@AllArgsConstructor
@NoArgsConstructor
@Builder
public class Product {

    private UUID id;

    private String title;

    private Double price;

    private Currency currency;

}