# app/src/main/java/io/swagger/dto/request/ProductRequest.java
package io.swagger.dto.request;

import io.swagger.enumeration.Currency;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@AllArgsConstructor
@NoArgsConstructor
@Builder
public class ProductRequest {

    private String title;

    private Double price;

    private Currency currency;

}